+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Optimizing performance for sync down operation.
+- Unloading synced records from online store after sync down operation.
 
 ## [0.7.1-beta.8] - 2016-11-20
 ### Fixed